MultiPackIndexV1: Calculate memsize with long
Errorprone pointed out that adding with int and casting to long at the
end could overflow.

Mark one of the operands as long to do the addition directly as long
and avoid the overflow risk.

+++ org.eclipse.jgit/src/org/eclipse/jgit/internal/storage/midx/MultiPackIndexV1.java
@@ -146,7 +146,7 @@ class MultiPackIndexV1 implements MultiPackIndex {
 		}
 
 		long getMemorySize() {
-			return byteArrayLengh(offsets) + byteArrayLengh(largeOffsets);
+			return (long)byteArrayLengh(offsets) + byteArrayLengh(largeOffsets);
 		}
 	}
 
@@ -227,7 +227,7 @@ class MultiPackIndexV1 implements MultiPackIndex {
 		}
 
 		long getMemorySize() {
-			return 4 + byteArrayLengh(oidLookup) + (FANOUT * 4);
+			return 4L + byteArrayLengh(oidLookup) + (FANOUT * 4);
 		}
 	}
 }
